login: fallback to ticket if password is not set
Our API backend allows using an existing ticket to create a new
ticket, as long as the original one is still valid.

For the recent work to enable auto-extending older tickets we added
some logic to temporarily set the ticket as password before calling
login, but we actually can just move the fallback into the login
method directly.

This safes some code and also allows users of this module to
explicitly trigger a re-login without having to care about what to use
as login credential secret.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>