api token: transparently add API token name in Authorization header
It's annoying that the client needs to know about how an API token
Authorization header needs to be formatted to be accepted by a PVE (or
PMG) API backend, so provide a more convenient way by introducing a
new api_token_name parameter that defaults to the one from PVE for
convenience.

With that one can just pass the full token (id+secret) and be done.
For backward compatibility we only add the name if it looks like there
isn't already one present.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>